lezione4/docker/python/requirements.txt
sqlalchemy
pg8000
paho.mqtt
